requirements.txt
openai>=1.35.9
python-dotenv>=1.0.1
pydantic>=2.7.0
matplotlib>=3.8.4
numpy>=1.26.4
typing_extensions>=4.9.0
langgraph>=0.2.34
langchain-core>=0.2.40
